+++ lib/features/dashboard/performance_table/domain/entities/get_custodian_performance_entity.dart
@@ -13,8 +13,8 @@ class GetCustodianPerformanceEntity extends Equatable {
     final String serialNumber;
     final String custodianName;
     final double performance;
-    final int amount;
-    final int riskPa;
+    final double amount;
+    final double riskPa;
     final double sharpeRatio;
